lower max width for comparison > spoiler for unit3d to 1000 instead of 1200

+++ src/bbcode.py
@@ -149,7 +149,7 @@ class BBCODE:
             comp_images = comp.split(']', 1)[1].replace('[/comparison]', '').replace(',', '\n').replace(' ', '\n')
             comp_images = re.findall("(https?:\/\/.*\.(?:png|jpg))", comp_images, flags=re.IGNORECASE)
             screens_per_line = len(comp_sources)
-            img_size = max_width / screens_per_line
+            img_size = int(max_width / screens_per_line)
             if img_size > 350:
                 img_size = 350
             for img in comp_images:
